+++ crates/holochain/src/conductor/api/api_external/app_interface.rs
@@ -109,12 +109,20 @@ impl AppInterfaceApi for RealAppInterfaceApi {
                     .await?;
                 Ok(AppResponse::CloneCellCreated(installed_clone_cell))
             }
-            AppRequest::ArchiveCloneCell(payload) => {
+            AppRequest::DisableCloneCell(payload) => {
                 self.conductor_handle
                     .clone()
-                    .archive_clone_cell(&*payload)
+                    .disable_clone_cell(&*payload)
                     .await?;
-                Ok(AppResponse::CloneCellArchived)
+                Ok(AppResponse::CloneCellDisabled)
+            }
+            AppRequest::EnableCloneCell(payload) => {
+                let enabled_cell = self
+                    .conductor_handle
+                    .clone()
+                    .enable_clone_cell(&*payload)
+                    .await?;
+                Ok(AppResponse::CloneCellEnabled(enabled_cell))
             }
             AppRequest::GossipInfo(payload) => {
                 let info = self.conductor_handle.gossip_info(&payload.dnas).await?;