+++ crates/holochain_conductor_api/src/admin_interface.rs
@@ -355,19 +355,12 @@ pub enum AdminRequest {
     /// [`AdminResponse::ZomeCallCapabilityGranted`]
     GrantZomeCallCapability(Box<GrantZomeCallCapabilityPayload>),
 
-    /// Restore a clone cell that was previously archived.
+    /// Delete a clone cell that was previously disabled.
     ///
     /// # Returns
     ///
-    /// [`AdminResponse::CloneCellRestored`]
-    RestoreCloneCell(Box<RestoreCloneCellPayload>),
-
-    /// Delete all clone cells that were previously archived.
-    ///
-    /// # Returns
-    ///
-    /// [`AdminResponse::ArchivedCloneCellsDeleted`]
-    DeleteArchivedCloneCells(Box<DeleteArchivedCloneCellsPayload>),
+    /// [`AdminResponse::CloneCellDeleted`]
+    DeleteCloneCell(Box<DeleteCloneCellPayload>),
 }
 
 /// Represents the possible responses to an [`AdminRequest`]
@@ -526,11 +519,8 @@ pub enum AdminResponse {
     /// The successful response to an [`AdminRequest::GrantZomeCallCapability`].
     ZomeCallCapabilityGranted,
 
-    // The successful response to an [`AdminRequest::RestoreCloneCell`].
-    CloneCellRestored(InstalledCell),
-
-    /// The successful response to an [`AdminRequest::DeleteArchivedCloneCells`].
-    ArchivedCloneCellsDeleted,
+    /// The successful response to an [`AdminRequest::DeleteCloneCell`].
+    CloneCellDeleted,
 }
 
 /// Error type that goes over the websocket wire.